pips: address satyam's comments
